chore: no chipper model in docker build (#168)
Previously, the Dockerfile was attempting to bake in the 2.6GB chipper model, but was no longer baking in the default detectron2 model. This was leading to errors when attempting to build the image in CI due to size issues. In addition, it's actually better to pre-load detectron2 in the first place rather than Chipper since it is the preferred "hi_res" model (as used to be the case).

This PR corrects that by only baking in the detectron2 image.

Since the chipper model is no longer baked into the image, there is the option of warming/downloading the chipper model when the application starts in scripts/app-start.sh. By default, it is not downloaded.*

*If it is not downloaded, it will be loaded on demand the first time it is requested. Generally, the download can take about a minute so that first request would be extra slow.