enhancement: add a top level exception handler to api (#224)
Add an error handler to the api to catch all exceptions. This will
standardize logging and formatting of error responses. If there's a ENV
variable set, log the stack in json, for better parsing in deployed
environments. Soon we'll have a general json logger, but this will tie
us over for now.

Handle HTTPError separate from the base Exception because Uvicorn
exceptions don't seem to get picked up under the base handler. Instead,
we can standardize on HTTPError by converting anything else that we
catch.

There are 2 * 2 cases to test out:

* A HTTPError is thrown / any other Exception is thrown
* The ENV var is one of `["dev", "prod"]` / No ENV is set

To get a HTTPError, send an unsupported file to the api:
```
curl -X POST 'http://localhost:8000/general/v0/general' --form files="@logger_config.yaml"
{"detail":"Unable to process logger_config.yaml: File type None is not supported."}%    
```

To test the generic handler, you can send a json file like so. This
throws a ValueError (which should be handled as a 400, tbd):
```
echo '{"hello": "world"}' > test.json

$ curl -X POST 'http://localhost:8000/general/v0/general' --form files="@test.json"                 
{"detail":"Detected a JSON file that does not conform to the Unstructured schema. partition_json currently only processes serialized Unstructured output."}%   
```

Try both of these cases and confirm the responses are consistent. Then,
run the server with `export ENV=prod` and confirm that the errors are
logged in one string.

Note - generic exceptions are still logged in the `uvicorn.error`
stream. Haven't figured out how to disable this, so some errors will be
dumped twice.