fix: fix local variable 'elements' referenced before assignment (#290)
We've seen this error in the hosted api:

```
UnboundLocalError: local variable 'elements' referenced before assignment
File "prepline_general/api/general.py", line 649, in response_generator
    response = pipeline_api(
  File "prepline_general/api/general.py", line 464, in pipeline_api
    for i, element in enumerate(elements):
```

The if block in the check for `zipfile.BadZipFile` allows us to fall
through without assigning `elements`. Instead, we should raise a 400
error when we catch this exception.