fix(chunking): minimal fix for combine arg typo (#345)
This is the minimal fix for the `combine_text_under_n_chars` arg-name
typo bug described in #337.

Chunking parameters were added to the API about 4 months ago in this
commit:
https://github.com/Unstructured-IO/unstructured-api/commit/49fe71026e026530276bbb72f6a9df46feb6d1c2

At that time, the identifier `combine_under_n_chars` was used instead of
`combine_text_under_n_chars` (as that parameter appears in
`chunk_by_title()`). As a result, any value supplied for this argument
via the REST API is ignored and defaults to 500. The
`combine_under_n_chars` parameter name appears in the API documentation
and elsewhere (README.md etc.) and differs from the parameter name
documented for `chunk_by_title()`.

A fuller remedy will be to add a separate `combine_text_under_n_chars`
parameter to the API endpoint such that either name is accepted for now,
then update the documentation such that all mentions are
`combine_text_under_n_chars`. At some point in the future we may remove
the `combine_under_n_chars` parameter from the endpoint.

However, making those changes requires a synchronized release of the
`unstructured` library and in the interest of time we're making this
"quick-fix" that can be released independently of any other
dependencies.