rfctr: let FastAPI do conversion of chunking args
FastAPI will do the type validation and conversion of the HTTP
form-fields we use for request arguments if we let it. Doing so gives us
better functionality (type validation of args) and eliminates the noisy
code we used to convert these from `List[str]` etc.

Change these for chunking options. If we like it we can change the rest
(recommended). Propagate the type changes through to `pipeline_api()`
and remove the arg parsing code.

Also remove setting default values for arguments to `pipeline_api()` for
chunking options. An internal call like this should avoid defaults and
the one caller (`pipeline_1()`) doesn't use them so they are dead code
and unwelcome noise.