fix: revert the removal of the custom openapi doc (#369)
We're now generating the proper openapi spec via FastAPI. The problem
is, if we point the Speakeasy clients to the new spec, the generated
code has too much churn. All good stuff if we were releasing the clients
for the first time, but now it will break compatibility. We'll need to
maintain a separate openapi spec for now, and add new params here
whenever they go into the server.

Example speakeasy churn:
https://github.com/Unstructured-IO/unstructured-python-client/pull/42