build: move `arm64` image to `wolfi-base` (#433)
### Summary

Updates the `arm64` image to use `wolfi-base` instead of `rockylinux`
and consolidates the `amd64` and `arm64` images into the same
Dockerfile. As of this PR, the `amd64` and `arm64` images for the API
are at parity.

### Testing

Successful docker build on the feature branch can be seen in [this
job](https://github.com/Unstructured-IO/unstructured-api/actions/runs/9875409234/job/27272072089).