feat: use the library to detect filetype
The mimetype detection has always been very naive in the API - we rely on the file extension. If the user doesn't include a filename, we return an error that `Filetype None is not supported`. The library has a `detect_filetype` that actually inspects the file bytes, so let's reuse this.