feat: Add content_type api parameter
Add an optional `content_type` param that allows the user to override the filetype detection. We'll use this value if it's set, or take the `file.content_type` which is based on the multipart `Content-Type` header. This provides an alternative when clients are unable to modify the header.

Also, rename a `content_type` variable that's actually referring to the response mimetype.